[Feature/DEV-5] Page Layout, Header 구현 및 배치 & News Ticker 배치 (#5)
* refactor: button 내 icon id 없을 경우 보이는 error text 변경

* design: add default svg size & color

* refactor: rename svg id, and use iconId type

* move: headline news ticker data & content load

* feat: create header

* move: news ticker container css
