+++ index.js
@@ -1,0 +1,28 @@
+import { createNewsTicker } from "./src/components/newsTicker/newsTicker.js";
+import { leftNewsItems, rightNewsItems } from "./src/data/headlineNews.js";
+
+/* render current time */
+document.addEventListener("DOMContentLoaded", () => {
+  const timeElement = document.getElementById("current-date");
+  const now = new Date();
+  const formattedDate = now.toLocaleDateString("ko-KR", {
+    year: "numeric",
+    month: "2-digit",
+    day: "2-digit",
+    weekday: "long",
+  });
+
+  timeElement.dateTime = now.toISOString();
+  timeElement.textContent = formattedDate;
+});
+
+/* render headline news ticker */
+document.addEventListener("DOMContentLoaded", () => {
+  const container = document.getElementById("news-ticker-container");
+
+  const left = createNewsTicker({ newsItems: leftNewsItems, tag: "연합뉴스" });
+  const right = createNewsTicker({ newsItems: rightNewsItems, tag: "연합뉴스" }, 1);
+
+  container.appendChild(left);
+  container.appendChild(right);
+});
