+++ src/components/button/button.css
@@ -30,7 +30,9 @@
 
 /* icon */
 .button > div {
-  height: 12px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
 }
 .button-icon {
   width: 12px;
