+++ src/components/newsTicker/newsTicker.css
@@ -1,8 +1,3 @@
-#news-ticker-container {
-  display: flex;
-  gap: 20px;
-}
-
 .news-ticker {
   background-color: var(--surface-alt);
   border: 1px solid var(--border-default);
@@ -21,7 +16,7 @@
 
   position: absolute;
   width: 100%;
-  transition: opacity 1s, transform 1s;
+  transition: opacity 0.2s, transform 1s;
 }
 
 .news-ticker p {
