+++ src/data/headlineNews.js
@@ -1,0 +1,15 @@
+export const leftNewsItems = [
+  { id: 1, title: "[1보] 김기현·안철수·천하람·황교안, 與전대 본경선 진출", link: "#" },
+  { id: 2, title: "뉴스 항목 2", link: "#" },
+  { id: 3, title: "뉴스 항목 3", link: "#" },
+  { id: 4, title: "뉴스 항목 4", link: "#" },
+  { id: 5, title: "뉴스 항목 5", link: "#" },
+];
+
+export const rightNewsItems = [
+  { id: 1, title: "뉴스 항목 1", link: "#" },
+  { id: 2, title: "뉴스 항목 2", link: "#" },
+  { id: 3, title: "뉴스 항목 3", link: "#" },
+  { id: 4, title: "뉴스 항목 4", link: "#" },
+  { id: 5, title: "뉴스 항목 5", link: "#" },
+];
