+++ src/features/unsubscribeDialog/unsubscribeDialog.css
@@ -1,3 +1,3 @@
-.dialog-button.cancle {
+.dialog-button.cancel {
   color: var(--text-strong);
 }
