+++ src/style/color.css
@@ -19,6 +19,7 @@
   --surface-default: #ffffff;
   --surface-brand-default: #4362d0;
 
+  --text-point: #4362d0;
   --text-default: #5f6e76;
   --text-white-default: #ffffff;
   --text-weak: #879298;
